+++ backend/app/models/__init__.py
@@ -1,0 +1,15 @@
+from app.models.notification_log import NotificationLogs
+from app.models.reminder import Reminders
+from app.models.subscription import Subscriptions
+from app.models.subscription_tag import SubscriptionTags
+from app.models.tag import Tags
+from app.models.user import User
+
+__all__ = [
+    "User",
+    "Subscriptions",
+    "Tags",
+    "SubscriptionTags",
+    "Reminders",
+    "NotificationLogs",
+]
