+++ backend/app/models/reminder.py
@@ -1,0 +1,17 @@
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer
+from sqlalchemy.sql import func
+
+from app.core.database import Base
+
+
+class Reminders(Base):
+    __tablename__ = "reminders"
+
+    id = Column(Integer, primary_key=True)
+    subscription_id = Column(
+        Integer, ForeignKey("subscriptions.id", ondelete="CASCADE"), nullable=False
+    )
+    remind_at = Column(DateTime(timezone=True), nullable=False)
+    is_recurring = Column(Boolean, default=True)
+    sent = Column(Boolean, default=False)
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
