+++ backend/app/models/tag.py
@@ -1,0 +1,14 @@
+from sqlalchemy import Column, ForeignKey, Integer, String, UniqueConstraint
+
+from app.core.database import Base
+
+
+class Tags(Base):
+    __tablename__ = "tags"
+    __table_args__ = (UniqueConstraint("user_id", "name", name="unique_user_tag"),)
+
+    id = Column(Integer, primary_key=True)
+    user_id = Column(
+        Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False
+    )
+    name = Column(String, nullable=False)
